+++ acme/resources/AE.py
@@ -49,8 +49,8 @@ class AE(AnnounceableResource):
 									 ])
 
 
-	def validate(self, originator:str=None, create:bool=False) -> Result:
-		if not (res := super().validate(originator, create)).status:
+	def validate(self, originator:str=None, create:bool=False, dct:dict=None) -> Result:
+		if not (res := super().validate(originator, create, dct)).status:
 			return res
 
 		self.normalizeURIAttribute('poa')