+++ acme/resources/CSEBase.py
@@ -53,8 +53,8 @@ class CSEBase(Resource):
 									 ])
 
 
-	def validate(self, originator:str=None, create:bool=False) -> Result:
-		if not (res := super().validate(originator, create)).status:
+	def validate(self, originator:str=None, create:bool=False, dct:dict=None) -> Result:
+		if not (res := super().validate(originator, create, dct)).status:
 			return res
 		
 		self.normalizeURIAttribute('poa')