+++ acme/resources/FCNT.py
@@ -100,8 +100,8 @@ class FCNT(AnnounceableResource):
 
 
 	# Checking the presence of cnd and calculating the size
-	def validate(self, originator:str=None, create:bool=False) -> Result:
-		if not (res := super().validate(originator, create)).status:
+	def validate(self, originator:str=None, create:bool=False, dct:dict=None) -> Result:
+		if not (res := super().validate(originator, create, dct)).status:
 			return res
 		return self._validateChildren(originator)