+++ acme/resources/RBO.py
@@ -35,8 +35,8 @@ class RBO(MgmtObj):
 	#	validate() and update()
 	#
 
-	def validate(self, originator: str = None, create: bool = False) -> Result:
-		if not (res := super().validate(originator, create)).status:
+	def validate(self, originator:str=None, create:bool=False, dct:dict=None) -> Result:
+		if not (res := super().validate(originator, create, dct)).status:
 			return res
 		self.setAttribute('rbo', False, overwrite=True)	# always set (back) to True
 		self.setAttribute('far', False, overwrite=True)	# always set (back) to True