+++ acme/resources/SUB.py
@@ -66,8 +66,8 @@ class SUB(Resource):
 		return CSE.notification.updateSubscription(self, newDict, previousNus, originator)
 
  
-	def validate(self, originator:str=None, create:bool=False) -> Result:
-		if (res := super().validate(originator, create)).status == False:
+	def validate(self, originator:str=None, create:bool=False, dct:dict=None) -> Result:
+		if (res := super().validate(originator, create, dct)).status == False:
 			return res
 		Logging.logDebug(f'Validating subscription: {self.ri}')